refactor: split TLB analysis into focused modules
